requirements.txt
streamlit>=1.28.0
langchain-openai>=0.1.0
langgraph>=0.0.40
python-dotenv>=1.0.0
httpx>=0.24.0
pandas>=2.0.0
